+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file is a history of the changes made to @idearium/cli.
 
 - Updated all commands to output an error and show the command help, if you try and run a command that doesn't exist.
 - `c dc rebuild <service>` will now always recreate containers.
+- Added the `--remove-orphans` to all `c dc down` command, as standard.
 
 ## 1.0.0-alpha.10